Notebooks/requirements.txt
langchain-core>=0.1.0
langchain-openai>=0.1.0
langgraph>=0.1.0
openai>=1.0.0
langchain>=0.1.0
langchain-community>=0.1.0
faiss-cpu>=1.7.0
numpy>=1.24.0
requests>=2.31.0
